+++ Makefile
@@ -30,16 +30,16 @@ run-daemon: test lint
 build-daemon: $(DAEMON_DIRS) $(DAEMON_FILES)
 	# make test
 	# make lint
-	@${PYTHON} -m nuitka daemon/__main__.py \
+	@${PYTHON} -m nuitka lento/daemon/main.py  \
 		--standalone
-	@mv __main__.dist/__main__ __main__.dist/lentodaemon
+	@mv __main__.dist/__main__.bin __main__.dist/lentodaemon
 	@sudo rm -rf __lentodaemon__.dist
 	@sudo cp -r __main__.dist __lentodaemon__.dist
 	@sudo rm -rf __main__.dist
 	@sudo rm -rf /usr/local/bin/__lentodaemon__.dist
 	@sudo cp -r __lentodaemon__.dist /usr/local/bin
 	@mkdir -p ~/Library/LaunchAgents
-	@sudo cp daemon/supporting_files/com.lento.lentodaemon.plist ~/Library/LaunchAgents
+	@sudo cp lento/daemon/supporting_files/com.lento.lentodaemon.plist ~/Library/LaunchAgents
 
 build-daemon-windows: test lint daemon/
 	@${PYTHON} -m nuitka daemon/__main__.py \
@@ -47,13 +47,13 @@ build-daemon-windows: test lint daemon/
 	@mv __main__.exe lentodaemon.lento.exe
 
 
-build-macos: build-daemon $(APP_DIRS) $(APP_FILES)
+build-macos: $(APP_DIRS) $(APP_FILES)
 	# make test
 	# make lint
 	@pyinstaller --name="Lento" \
 		--add-data "lento.qss:." \
 		--add-data "fonts/*.ttf:fonts/" \
-		--add-data ".venv/lib/python3.10/site-packages/fleep/data.json:fleep/." \
+		--add-data ".venv/lib/python3.11/site-packages/fleep/data.json:fleep/." \
 		--add-data "assets/toggle-unfolded.svg:assets/." \
 		--add-data "assets/toggle-folded.svg:assets/." \
 		--add-data "assets/arrow-left.svg:assets/." \
