+++ daemon.py
@@ -1,0 +1,3 @@
+from lento.daemon import main
+
+main.main()
