+++ lento/daemon/lento_blocker_plugin.py
@@ -1,13 +1,13 @@
 import logging
 from typing import Optional
-from daemon.util.db import DBController
-from daemon.util.util import format_website
-from daemon.alert.notifications_controller import NotifsController
-from daemon.alert.notifications_controller import (
+from lento.daemon.util.db import DBController
+from lento.daemon.util.util import format_website
+from lento.daemon.alert.notifications_controller import NotifsController
+from lento.daemon.alert.notifications_controller import (
     WEBSITE_DEFAULT_TITLE,
     WEBSITE_INFO_DEFAULT_MSG,
 )  # noqa: E501
-from daemon.alert.notifications_controller import WEBSITE_CONFIRM_DEFAULT_MSG
+from lento.daemon.alert.notifications_controller import WEBSITE_CONFIRM_DEFAULT_MSG
 from proxy.http import httpStatusCodes
 from proxy.common.flag import flags
 from proxy.http.proxy import HttpProxyBasePlugin
