+++ lento/daemon/util/util.py
