+++ lento/desktop_client/icon_manager.py
