+++ lento/desktop_client/viewcontrollers/block_item_details.py
@@ -1,7 +1,7 @@
 import copy
 import logging
 import platform
-from lento import utils
+from lento.desktop_client import utils
 from PySide6.QtGui import QIcon, Qt, QPixmap
 from PySide6.QtWidgets import QFileDialog, QSpinBox
 from PySide6.QtWidgets import (
@@ -11,10 +11,10 @@ from PySide6.QtWidgets import (
     QVBoxLayout,
     QPushButton,
 )  # noqa: E501
-from lento.views.textedit import LentoTextEdit
-from lento.viewcontrollers.popup_list import LentoPopUpList
-from lento.model.block_items import LentoAppItem, LentoWebsiteItem
-from lento.views.toggle import Toggle
+from lento.desktop_client.views.textedit import LentoTextEdit
+from lento.desktop_client.viewcontrollers.popup_list import LentoPopUpList
+from lento.desktop_client.model.block_items import LentoAppItem, LentoWebsiteItem
+from lento.desktop_client.views.toggle import Toggle
 
 
 class LentoBlockItemDetailsViewMode:
