+++ lento/desktop_client/views/button.py
@@ -1,4 +1,4 @@
-from lento import utils
+from lento.desktop_client import utils
 from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import QPushButton
