+++ lento/desktop_client/views/resizable_stack.py
