+++ lento/desktop_client/views/resizing_tab.py
