+++ lento/desktop_client/views/textedit.py
