+++ lento/desktop_client/views/toggle_list.py
@@ -1,8 +1,8 @@
 import logging
-from lento import utils
+from lento.desktop_client import utils
 from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import QPushButton, QVBoxLayout, QWidget
-from lento.views.list_items import LentoListItem
+from lento.desktop_client.views.list_items import LentoListItem
 
 
 class LentoToggleList(QWidget):
